+++ crates/core/src/view/mod.rs
@@ -105,6 +105,7 @@ pub mod clock;
 pub mod common;
 pub mod cover_editor;
 pub mod about_dialog;
+pub mod cloud_dialog;
 pub mod dialog;
 pub mod dictionary;
 pub mod email_dialog;